% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_data.R
\docType{data}
\name{epa_data}
\alias{epa_data}
\title{EPA Fuel Economy and other data for cars sold in the US from 1984 through 2021}
\format{
A data frame with 42426 rows and 9 variables:
\describe{
\item{city}{EPA measured fuel economy in miles per gallon}
\item{hwy}{EPA measured fuel economy in miles per gallon}
\item{cyl}{The number of cylinders in the engine}
\item{disp}{The volumetric displacement of the engine in litres}
\item{drive}{The vehicle drivetrain layout}
\item{make}{The name of the vehicle manufacturer}
\item{model}{The vehicle model name}
\item{trans}{The transmission type, either manual or automatic}
\item{year}{The vehicle model year}

...
}
}
\source{
\url{https://www.fueleconomy.gov/feg/download.shtml}
}
\usage{
epa_data
}
\description{
A data set containing data for over 42,000 cars sold in the US from 1984 through 2021
}
\keyword{datasets}
